requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
google-generativeai==0.3.1